requirements.txt
python-dotenv
openai>=1.0.0
Pillow>=9.0.0
requests>=2.31.0
python-magic>=0.4.27 